rename: AllClearButton 오타 수정
