+++ src/components/AllClearButton.js
